Connor's comments
