+++ packages/agent-ui/src/actions/notifications.js
@@ -1,4 +1,3 @@
-import { v4 as uuid } from 'uuid';
 import * as actionTypes from '../constants/actionTypes';
 
 export const addNotifications = data => ({
@@ -10,14 +9,3 @@ export const removeNotifications = keys => ({
   type: actionTypes.REMOVE_NOTIFICATIONS,
   keys
 });
-
-export const makeNotification = args => ({
-  key: uuid(),
-  ...args
-});
-
-export const makeNewSegmentNotification = (agent, process, segment) => {
-  // extract mapId and linkHash
-  const { link: { meta: { mapId } }, meta: { linkHash } } = segment;
-  return makeNotification({ agent, process, mapId, linkHash });
-};
