+++ packages/agent-ui/src/actions/notifications.test.js
@@ -1,11 +1,5 @@
-import 'babel-polyfill';
 import { expect } from 'chai';
-import {
-  addNotifications,
-  removeNotifications,
-  makeNotification,
-  makeNewSegmentNotification
-} from './notifications';
+import { addNotifications, removeNotifications } from './notifications';
 import * as actionTypes from '../constants/actionTypes';
 
 describe('notifications actions', () => {
@@ -25,33 +19,5 @@ describe('notifications actions', () => {
       type: actionTypes.REMOVE_NOTIFICATIONS,
       keys
     });
-  });
-});
-
-describe('notification helpers', () => {
-  it('makeNotification() generates a new notification object with uuid', () => {
-    const obj = { foo: 'bar', one: 2 };
-    const { key, ...fromObj } = makeNotification(obj);
-    expect(key).to.have.lengthOf(36);
-    expect(fromObj).to.deep.equal(obj);
-  });
-
-  it('makeNewSegmentNotification() adds agent info', () => {
-    const agent = 'foo';
-    const process = 'bar';
-    const mapId = 'foo/bar';
-    const linkHash = 'xyz';
-    const obj = {
-      agent,
-      process,
-      segment: {
-        link: { meta: { mapId } },
-        meta: { linkHash }
-      }
-    };
-    const { key, ...fromObj } = makeNewSegmentNotification(
-      ...Object.values(obj)
-    );
-    expect(fromObj).to.deep.equal({ agent, process, mapId, linkHash });
   });
 });
